third_party/blink/web_tests/http/tests/websocket/echo-with-no-extension-expected.txt

Test WebSocket.extensions attribute with no extension support in server side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check the value before opening the connection:
PASS ws.extensions is ""
Check if the value is read only:
PASS ws.extensions = 'foo'; ws.extensions is ""
Check the value after the connection is established:
PASS ws.extensions is ""
PASS successfullyParsed is true

TEST COMPLETE
